+++ index.html
@@ -3,16 +3,13 @@
   <head>
     <meta charset="UTF-8" />
     <title>weebsync</title>
-      <link rel="stylesheet" href="./src/renderer/main.scss">
+      <link rel="stylesheet" href="./src/client/main.scss">
       <link rel="stylesheet" href="./node_modules/vuetify/dist/vuetify.min.css">
-      <link rel="stylesheet" href="./node_modules/vue2-perfect-scrollbar/dist/vue2-perfect-scrollbar.css">
-    <meta
-      http-equiv="Content-Security-Policy"
-      content="default-src 'self'; script-src 'self'; style-src 'self' 'nonce-dQw4w9WgXcQ'"
-    />
+      <link rel="stylesheet" href="./node_modules/vue3-perfect-scrollbar/dist/vue3-perfect-scrollbar.css">
+
   </head>
   <body>
     <div id="app"></div>
-  <script type="module" src="./src/renderer/index.ts"></script>
+  <script type="module" src="./src/client/index.ts"></script>
   </body>
 </html>
