+++ package.json
@@ -5,14 +5,10 @@
   "license": "MIT",
   "main": "./build/main.js",
   "scripts": {
-    "build-main-preload": "node build.js",
-    "build-renderer": "vite build",
+    "client:build": "vite build",
+    "client:dev": "vite dev",
     "build": "yarn run build-renderer && yarn run build-main-preload",
-    "watch": "tsc -w",
-    "lint": "eslint -c .eslintrc --ext .ts ./src",
-    "start": "npm run build && electron ./build/main.js",
-    "app:dist": "yarn build && electron-builder -p never --win",
-    "postinstall": "electron-builder install-app-deps"
+    "lint": "eslint -c .eslintrc --ext .ts ./src"
   },
   "build": {
     "appId": "weeb.sync",
@@ -32,38 +28,40 @@
   },
   "author": "Bastian Ganze",
   "devDependencies": {
-    "@types/node": "^17.0.8",
+    "@types/node": "^20.4.0",
     "@types/progress-stream": "^2.0.2",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "electron": "22.0.2",
-    "electron-builder": "^23.6.0",
-    "esbuild": "^0.14.11",
-    "eslint": "^7.32.0",
-    "prettier": "^2.8.3",
-    "sass": "1.32.13",
-    "sass-loader": "^10.1.1",
-    "typescript": "4.5",
-    "vue": "^2.6.12",
+    "@typescript-eslint/eslint-plugin": "^5.61.0",
+    "@typescript-eslint/parser": "^5.61.0",
+    "esbuild": "^0.18.11",
+    "eslint": "^8.44.0",
+    "eslint-plugin-vue": "^9.15.1",
+    "prettier": "^3.0.0",
+    "sass": "1.63.6",
+    "sass-loader": "^13.3.2",
+    "typescript": "5.1.6",
+    "vue": "^3.3.4",
     "vue-class-component": "^7.2.6",
-    "vue-property-decorator": "^9.1.2",
-    "vue-router": "^3.5.1",
-    "vuex": "^3.6.2"
+    "vue-eslint-parser": "^9.3.1",
+    "vue-property-decorator": "^9.1.2"
   },
   "dependencies": {
-    "@fontsource/lato": "^4.5.0",
+    "@fontsource/lato": "^5.0.4",
     "@mdi/font": "^7.1.96",
     "@mdi/js": "^7.1.96",
+    "@vitejs/plugin-vue": "^4.2.3",
     "basic-ftp": "^5.0.2",
     "chokidar": "^3.5.2",
     "handlebars": "^4.7.7",
-    "strongly-typed-events": "^2.1.14",
-    "ts-pattern": "^4.1.3",
-    "vite": "^2.0.2",
-    "vite-plugin-components": "^0.13.2",
-    "vite-plugin-vue2": "^1.2.1",
-    "vue-template-compiler": "^2.6.12",
-    "vue2-perfect-scrollbar": "^1.5.2",
-    "vuetify": "^2.6.2"
+    "pinia": "^2.1.4",
+    "socket.io": "^4.7.1",
+    "socket.io-client": "^4.7.1",
+    "strongly-typed-events": "^3.0.9",
+    "ts-pattern": "^5.0.1",
+    "unplugin-vue-components": "^0.25.1",
+    "vite": "^4.4.2",
+    "vite-plugin-compression": "^0.5.1",
+    "vite-plugin-vuetify": "^1.0.2",
+    "vue3-perfect-scrollbar": "^1.6.1",
+    "vuetify": "^3.3.7"
   }
 }
