+++ src/client/plugins/pinia.ts
@@ -1,0 +1,3 @@
+import {createPinia} from "pinia";
+
+export const pinia = createPinia();
