+++ src/client/plugins/vuetify.ts
@@ -1,0 +1,20 @@
+import "@mdi/font/css/materialdesignicons.css";
+import { createVuetify, VuetifyOptions } from "vuetify";
+
+const opts: Partial<VuetifyOptions> = {
+  icons: {
+    defaultSet: "mdiSvg",
+  },
+  theme: {
+    themes: {
+      dark: {
+        dark: true,
+        colors: {
+          primary: "#ffffff",
+        },
+      },
+    },
+  },
+};
+
+export default createVuetify(opts);
