--- src/renderer/shims-tsx.d.ts
@@ -1,14 +1,0 @@
-import Vue, { VNode } from "vue";
-
-declare global {
-  namespace JSX {
-    // eslint-disable-next-line @typescript-eslint/no-empty-interface
-    type Element = VNode;
-    // eslint-disable-next-line @typescript-eslint/no-empty-interface
-    type ElementClass = Vue;
-    interface IntrinsicElements {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      [elem: string]: any;
-    }
-  }
-}
